requirements.txt
absl-py==2.1.0
bchlib==2.1.2
focal-frequency-loss==0.3.0
kornia==0.7.3
kornia_rs==0.1.5
lightning-utilities==0.11.7
lpips==0.1.4
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
pillow==10.4.0
scipy==1.14.1
six==1.16.0
tensorboard==2.17.1
torch==2.4.1
torchmetrics==1.4.1
torchvision==0.19.1
timm
convnext-unet